Fix pan-zoom controls when leaving the canvas (#285)
### Summary
In `PanZoomControls` this change now uses pointer events instead of
mouse events and treats the `pointercancel` event in the same way as a
`pointerup` event, which prevents the surprising and undesirable
behavior described in #284. Another benefit of using pointer events is
that we can capture and release the pointer event ID, which allows us to
continue responding to pointer events when the pointer is down but
strictly outside the target.

### Related Issue
Closes #284

### Tests & Checks
I manually tested an example with pan-zoom controls and the main React
component.

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>