Handle upper case dimension names in OME-Zarr chunk loader (#288)
### Summary
In the OME-Zarr chunk loading parts of the code, we assume that the last
two spatial dimensions that we present will be named "x" or "y". We
mostly do this in a case-insensitive way, but `getDimInfoMap` currently
assumes lower-case only. See the image2d_from_omezarr5d_u8 example for a
case where this currently fails.

This change is a short term fix to that method to be case-insensitive in
a similar way to related parts of the code.

For a long term fix, we'll want a more robust way of handling dimension
names and correspondences, but that is a much larger problem.

### Tests & Checks
I manually tested all the examples after this change including the
currently broken image2d_from_omezarr5d_u8.

---------

Co-authored-by: phoenixAja <plogan@chanzuckerberg.com>