Add a ScaleBar React component to OmeZarrImageViewer (#301)
### Summary
This defines an internal ScaleBar component in the Idetik React package
and uses it in the OmeZarrImageViewer component. The scale bar's
visibility and alignment can be configured using the viewer's props. By
default it is visible (because an OME-Zarr often has a useful units) and
left-aligned to match the general layout.

I needed to reconfigure the div hierarchy in the OmeZarrImageViewer
component to handle resizing somewhat sensibly, especially at small
widths (e.g. mobile devices). I decided to place the scale bar in the
bottom-left corner for now, because it seemed like the least disruptive
option that was also good enough for the OrganelleBox portal launch.

I also allow units to be optionally returned when loading an OME-Zarr's
attributes. In the React component, these units are fetched and updated
when fetch the z-axis metadata.


https://github.com/user-attachments/assets/b359a5ce-c540-4e34-bdd9-9b4d5737f19e

### Related Issue
Closes #191

### Tests & Checks
I manually tested this with the main component and asserted that the
described length roughly matches the pixel sizing/spacing described in
the corresponding OME-Zarr file. I also manually created a more exact
test using a Python script to create and then serve an OME-Zarr with
integer pixel spacing.