Use canvas.clientWidth instead of idetik.width for scale-bar updates (#307)
### Summary
This fixes an issue with the reported physical length of the scale bar
when switching between displays with different DPIs (see #306) by using
`idetik.canvas.clientWidth` instead of `idetik.width` and
`window.devicePixelRatio` to handle the scale-bar length calculations.

Using the CSS width rather than the canvas buffer width is conceptually
simpler, and is also immune to Safari's partial support for
`window.devicePixelRatio`.

### Related Issue
Closes #306 

### Tests & Checks
I manually tested the scale bar example and the React component example
with the setup described in #306. The bug is present without these
changes and not present after. After these changes, I verified that the
reported physical length of a single pixel and the whole image is as
expected based on the image metadata.