Enable local directory source in <OmeZarrImageViewer> (#321)
I originally made `sourceUrl` and `sourceLocalDirectory` a "one of"
using a discriminated union, but realized that when using the local file
system, initially neither will be set because the user has to select the
directory first. So the component has been updated to support an initial
state with no source (hide all the stuff on the edges).

Diff looks better when hiding whitespace changes.

Maybe the demo in the base PR doesn't need to exist now that there's a
demo in react/?


https://github.com/user-attachments/assets/e542cbb9-d3c4-43a4-b65b-dc5644c1ff52