Add an explicit colormap to the label image layer  (#322)
### Summary
This change adds a colormap to the label image layer, so that certain
label values can be mapped to specific colors. This map is passed to the
fragment shader as a uint32 scalar texture where the label keys are in
the first row and the RGBA colors are packed into the second row. The
fragment shader iterates over those keys and unpacks the color when
needed.

It modifies an existing example showing multi-label segmentation image
overlaid on a corresponding intensity image from public OrganelleBox
data, with a specific segmented mapped to a green color outside the
other cycled colors.

<img width="1914" height="1132" alt="Screenshot 2025-07-24 at 11 02 51"
src="https://github.com/user-attachments/assets/5140c209-9ba7-437c-ab06-d50c1a50d29b"
/>

### Related Issue
Related to #177 

### Tests & Checks
I manually tested the new example, which works as expected.
I also manually tested the existing examples to check for breakage.