Add a label image series layer (#347)
### Summary
This introduces a `LabelImageSeriesLayer` to handle loading and
rendering label image data in a series across a chosen dimension.
Because this new layer overlaps with both the `LabelImageLayer` and
`ImageSeriesLayer`, I did some refactoring to limit large chunks of
repeated code, though there is still some remaining boiler plate.

The larger refactor is a new non-exported `ImageSeriesLoader` that
attempts to encapsulate the complex loading logic of the existing
`ImageSeriesLayer`, which is identical in the `LabelImageSeriesLoader`.
I expect this class to go away in the future (e.g. with a more
sophisticated `ChunkManager`), and I hope its introduction will make
that removal easier.

I also refactored the color-cycle and lookup-table into a single
`LabelColorMap` class that can map from label image pixel values to
colors (though that actually happens in the fragment shader).

Finally, I added a new example that is the same as the label image
overlay, but introduces a slider over the time dimension.

https://github.com/user-attachments/assets/999fa362-acbe-4716-9ccd-73e6ffd6c8e5

### Related Issue
Closes #177

### Tests & Checks
I manually tested the new example.
I manually tested the existing examples.
I manually tested the main React component.