Fix chunk visibility bounds check (#357)
### Summary
This fixes the visibility bounds check in the chunk manager, as well as
a bug in how we handle translations in OME-Zarr metadata and the
ultimate conversion to integer indices for an array at some multiscale
level.

### Related Issue
Closes #355 

### Tests & Checks
I tested the existing examples to ensure they continue to work as
expected.
I modified one of the existing examples to use OME-Zarr data that
includes a translation to test that this fix also works when the
underlying data contains an offset.