[small] Add support for color + layer transparency in Points renderable (#363)
### Summary
This adds basic support for transparency in the `Points` renderable
object. The shader also includes `u_opacity`, which corresponds to a
layer-level opacity value. This also allows for alpha-based
anti-aliasing depending on how the marker atlas is drawn or sampled. For
any of this to work, a layer using this will need to set its
`transparent` attribute to `true`.

### Related Issue
N/A

### Tests & Checks
Tested primarily in the Try CryoLens app, but the points example here
still works.