Move Z-plane slicing to the layer (#385)
Summary:
This PR modifies the chunk loading logic by storing the entire data set for
the chunk in the chunk data structure and moving the Z-plane slicing from the
loader to the image layer. I also refactored the loader now that I understand
the schema a little better by removing getDimInfoMap and writing a cleaner
helper.