Rename and move OME-Zarr and Zarr modules and types (#388)
### Summary
This change renames existing modules and types to prefer OME-Zarr over
OME-NGFF. Before this change, we used both names interchangeably which
could be confusing. I decided to use OME-Zarr because it is more
specific and consistent with our usage (e.g. `loadOmeZarrImage` assumes
the URL points to a Zarr group).

Some of these naming concerns came from comments in #376.

### Related Issue
Closes #366 

### Tests & Checks
I manually tested the examples and React components, which work as they
do on the main branch.