Refactor dimension mapping in chunk manager (#387)
### Summary
This is mostly a refactor that attempts to clarify and fix the dimension
mapping in the chunk manager and the OME-Zarr loader. It also brings
some validation and mapping on construction of the chunk manager source,
which allows us to skip some checks later.

### Tests & Checks
I manually tested all the examples that use the chunk manager.