Define distinct ChunkedImageLayer for chunk streaming (#403)
### Summary
This splits the existing image layer into two classes.

1. `ImageLayer`: this encapsulates the pre-chunk-streaming image layer,
which is crude but simple. This is currently used by the main React
component and the applications that use that.

2. `ChunkedImageLayer`: this uses chunk-streaming for a better
experience and performance, as well as built-in support for image
series, but at the cost of some complexity. This does not support some
features of the `ImageLayer` (e.g. multiple channels). This will
eventually replace at least the `ImageLayer` and `ImageSeriesLayer`.

The main motivation of this change is clarity. Since `ChunkedImageLayer`
has a very different loading pattern, the existing image layer had
become hard to read and maintain. This also allows us to move more
quickly on developing chunk streaming without breaking existing usage of
older layers.

Going forward, we intend to leave `ImageLayer` and `ImageSeriesLayer`
alone except for important maintenance, and will eventually delete them
in favor of `ChunkedImageLayer`. At the point, we may want to rename
`ChunkedImageLayer`.

I updated all the examples that I'm aware of using chunk streaming to
use the `ChunkedImageLayer`.

### Tests & Checks
I manually tested the existing examples and React components.