Remove ultrack-prototype workspace (#410)
### Summary
This removes the ultrack-prototype workspace.

Our other applications all live in other repos, which will be a more
maintainable model for us as we move towards releasing and versioning
Idetik.

I pushed two tags to make it easier to find this code if/when we need it
in the future:
*
**[last-ultrack-prototype](https://github.com/chanzuckerberg/idetik/releases/tag/last-ultrack-prototype)**
- Last commit with ultrack-prototype
*
**[removed-ultrack-prototype](https://github.com/chanzuckerberg/idetik/releases/tag/removed-ultrack-prototype)**
- Commit removing ultrack-prototype workspace

Now that I think about it, searching for this PR is probably the
easiest. ¯\\\_(ツ)\_/¯

### Related Issue
N/A

### Tests & Checks
Project still builds, no remaining references to the ultrack prototype
in the repo