Update React example to use OrganelleBox time metadata (#415)
### Summary
This change updates the React example to use the newly added time
metadata that is planned to be added in the OrganelleBox.

https://github.com/user-attachments/assets/a016746c-4e72-414b-948f-982b586c5372

The motivation behind this change was to use those metadata and share
with the OrganelleBox team to ensure this type of display is sufficient.
The example should also serve as an integration example.

As part of these changes, I modified the name and return type of
`loadOmeroDefaultZ`, so that we can also pull out `defaultT` as needed
for this example. I updated the `initialIndex` prop of the
`OmeZarrImageViewer` appropriately. I'd like to improve the state of
accessing these metadata in the future (e.g. avoiding the need for
loading/parsing the main OME-Zarr JSON multiple times), but did not
attempt that in this change.

### Tests & Checks
I manually tested the React examples.
I manually tested the manual examples.