Dynamic slice coordinates and a fixed dimension map for chunked images (#400)
### Summary
This change further splits the layers and data structures used for chunk
streaming loading, away from the older/cruder approach of loading a
single fixed region.

The main motivation behind this change is to separate the concerns of
`DimensionMap`, which currently maps from Idetik dimensions to OME-Zarr
array indices and also stores slicing coordinates.

The approach of this change is to define a new `SliceCoordinates` type
that optionally stores the z, c, and t coordinates that should be used
to index a particular point or index in each of those dimensions. This
effectively replaces the use of `Region` in the examples that use chunk
streaming.

Then also define a `SourceDimensionMap` that handles both the dimension
mapping and loader attributes, and is intended to be immutable.

### Tests & Checks
I manually tested the existing examples.
I manually tested the existing components.

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>