Implement `ChannelsEnabled` for `ChunkedImageLayer` (#422)
### Summary
Our current method for integrating with channel controls relies on
implementing the `ChannelsEnabled` interface. `ChunkedImageLayer`
currently only supports a single-channel (see #372), but implementing
this interface allows us to add dynamic contrast limit (window/level)
controls for these layers. This is needed for porting some existing
applications (Try Cryolens, Tomogram Quality) to use the
`ChunkedImagelayer` for progressive chunk streaming.

It's a little weird to have to pass a list of channel props when the
limit is 1. So we may want to change the interface later if we decide to
only support single-channel images in this layer (relying instead on
layer blending for multiple channels). I think this can be revisited
when we are ready to remove `ImageLayer` and `ImageSeriesLayer`.

Tagging @phoenixAja for review, since I think this is needed for porting
the Tomogram Quality app to use `ChunkedImageLayer`.

### Related Issue
#372 (does not close)

### Tests & Checks
Tested with current examples
Added window/level controls to chunk streaming example

---------

Co-authored-by: Phoenix Logan <plogan@chanzuckerberg.com>