Use preventDefault to prevent page scroll on zoom (#445)
### Summary
When the Idetik canvas is within a page that 

### Related Issue
Closes #440 

### Tests & Checks
I temporarily manually modified navigation.css to use `150%` height for
the body, which adds a vertical scrollbar. I then used the HCS example
as the canvas does not cover the whole page in that case, so I can zoom
in/out with the mouse wheel on the canvas, and scroll the page up/down
on the left side.