Horizontally grow the series slider in the OmeZarrImageViewer react component (#446)
### Summary
This reverts some of the changes from #397 to make the scale-bar and
load-button/slider live in the same div element at the bottom of the
canvas. This allows the slider to grow to horizontally fill the space at
the bottom of the canvas, when the canvas is small enough.

In order to allow pointer events (e.g. for pan/zoom control) through to
the canvas when there is empty space in that new div element, I used
`pointer-events: none` on that parent, then selectively applied
`pointer-events: auto` to descendants that need those events (i.e. the
button and slider). Unlike main, this also allows pan/zoom over the
scale-bar and series label, which seems desirable.

### Related Issue
Closes #441

### Tests & Checks
I manually tested the main React example before and after this change.