Support time in chunk manager and chunked image layer (#432)
### Summary
This change provides basic support for the time dimension in the chunk
manager, with no prefetching.

The biggest change here is that `ChunkManagerSource.chunks_` is now a 2D
array instead of 1D. The outer dimension represents time, so that we can
efficiently update only the chunks at the current or previous time
point.

The chunk manager source only works under some constraints about the
time dimension in the chunk source, and will throw on construction if
those are not met.

### Related Issue
Closes #453 

### Tests & Checks
I manually tested my changes in the examples that use the chunk manager.