Skip chunk manager update when not needed (#459)
### Summary
This skips `ChunkManager.update` when there are no sources. This avoids
a small amount of unnecessary computation (e.g. some small matrix
inversions and multiplications) on the CPU, but in a very simple way.

It's possible that all layers may use the `ChunkManager` in the future,
but this can be removed when that's the case.

### Related Issue
Closes #313

### Tests & Checks
I manually tested all the core examples.
I manually tested the main React component.