Dispatch `getChunk` to WebWorker pool (#463)
### Summary
This introduces a new `getChunk` replacement that dispatches the fetch +
(more importantly) chunk decompression to a pool of WebWorkers. The goal
was to make the replacement function close to a drop-in to minimize
changes to existing code. This is now working for both the chunk
streaming (zebrafish) and OME-Zarr v0.5 (marmoset neurons) examples. I
think it's ready for another set of eyes.

There are some slightly complicated aspects here, mostly around what can
types of objects can be transferred or copied to a WebWorker:
* use `ArrayBuffer` for zero-copy data transfer from the WebWorker back
to the main thread (see [MDN Transferable
Objects](https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Transferable_objects))
* need to re-open Zarr arrays separately in the WebWorker, so there is a
need to serialize all the required parameters - for this I added a
`ZarrArrayParams` that supports the types of arrays we currently
support.
* you can't pass through an `AbortController`/`AbortSignal` directly, so
this PR sets up a listener for the signal on the main thread, and sends
a `cancel` message to the worker pool, which creates/maintains its own
separate `AbortController` for each task.

Eventually we may want to make this more generic and integrate it with
the `ChunkQueue` to run any (or all) chunk-fetching operations in a
worker pool. I need to think a bit more about that design. An
intermediate step might be to make the worker _pool_ more generic, but
keep the zarr-specific workers. None of this feels urgent until we start
working with other types/sources of chunked data.

### Related Issue
Closes #456

### Tests & Checks
Tested with existing examples - framerate on chunk_streaming example
seems much improved when zooming/panning (loading new chunks)

I also tested by temporarily adding two features:
1. an option to force `getChunk` on the main thread instead of the
worker
2. a blocking sleep to simulate additional CPU bottleneck

With these changes it was very obvious the WebWorker was effective at
keeping the main thread unblocked.

Here are some videos comparing the frame rate of the chunk streaming
example on `main` vs. this branch. For this test I kept the browser
cache enabled, because it exacerbates the problem when decompression is
not naturally spaced out by network transfers.

### `main`


https://github.com/user-attachments/assets/1b5a2bbf-662b-4e73-9e74-1291b6a3edbb


### `getChunk` in WebWorkers (this branch)


https://github.com/user-attachments/assets/a8231cf5-6877-45b5-882d-585104e13b55

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>