Prepare for latest npm package versions (#468)
### Summary
This change is the result of running `npm version major` in both the
core and react packages.

I decided to bump the major version because this is a pre-release. I'm
not sure if there are breaking changes in either package and very few
apps depend on either of these packages, but this offers some protection
for them.

### Tests & Checks
I ran `npm install` to check that nothing changes after these bumps.
Then I ran `npm run build:all` to check that both packages build.
Then I ran `npm run examples` to check that the core examples work.
Then I ran `npm run components` to check that the react example works.