Prepare for npm prerelease version 4 (#471)
### Summary
This change was created by running `npm version major` on both the core
and react packages. I decided to bump the react version twice, so that
the pre-release version numbers for both packages are in sync for now.

I also removed most of the content from the READMEs since some of it is
already inaccurate, and I'd rather not encourage people to use Idetik if
they stumble across it!

### Tests & Checks
I tested this with the organelle-box portal, by building locally on this
branch, updating the yalc packages locally, and running a local dev
build of organelle-box.