Add event dispatcher connect/disconnect in runtime start/stop (#474)
### Summary
I attempted to update Idetik in the VCP project in preparation of the
CryoLens release. Unfortunately this broke the HITL app, which is now
tied to the same Idetik dependency version. The previous version did not
include #394, and the HITL app behavior changed without long-lived
Idetik runtime(s).

I was able to work around some of the issues there, but duplicate event
observers are still interfering. I was able to test a local build with
these changes and it fixes some of the issues.