Stop observing events when runtime stops (#473)
### Summary
This stops observing events and size changes when `Idetik.stop` is
called. It also refactors the size observation logic into a utility
class to improve the readability of the Idetik runtime. I also adding
some added some extra logs to help catch double starts or unnecessary
stops.

### Tests & Checks
I manually tested the changes with the core examples and the react
components.

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>