Relax time translation and scale constraint in chunk manager (#489)
### Summary
This change relaxes the constraint of a zero translation and unitary
scale in the time dimension in the chunk manager. It also adds an
example from the public OME-Zarr dataset that has a non-unitary scale.

While relaxing this constraint is generally useful for future work, I
prioritized this because many of the DCA datasets have a time scale.

### Tests & Checks
I manually tested the new example.
I manually tested the existing examples that use chunk streaming.