Basic support for channel dimension in chunk manager source (#493)
### Summary
This adds some support for the channel dimension in the chunk manager
source. The support requires that the there must be one channel per
chunk, as well as some other sensible constraints for the channel
dimension that are validated on construction.

This approach requires minimal changes within the chunk manager source,
and no changes in any dependencies of it. The main downside of the
approach is that incurs a performance penalty in
`updateChunksAtTimeIndex` that is linear in the number of channels. This
could be avoided by making changes to certain data structures (e.g.
storing the channel index in a new dimension in
`ChunkManagerSource.chunks_`, changing the `Chunk` type), but these
would likely have a relatively larger impact on dependencies.

The motivation for this change is twofold.

1. Split some of the larger work around support multi-channel chunk
streaming: https://github.com/chanzuckerberg/idetik/pull/492
2. Unblock work attempting to split the chunk manager source as part of
larger redesign/refactor work.

### Tests & Checks
I manually tested the core examples that use chunk streaming.