Introduce multiple-viewport API (breaking changes) (#506)
### Summary

This introduces an API for instantiating a multi-viewport runtime. The
bulk of the important changes are in `idetik.ts`. The rest of the
changes (primarily in the examples) are pretty mechanical, but useful
for seeing what the updated API looks like.

In the interest of keeping the configuration explicit, and validation
simple, this is a _breaking change_, hence all the updated examples. In
summary:
* When instantiating the runtime, you _must_ specify the `viewports`
array
* For single-viewport applications, simply specify a single viewport in
the array
* Each viewport in the config requires an HTML element that defines its
bounds, if you omit this it will default to the base `canvas` element
  * In general this is what you want for a single-viewport app
* Viewports cannot share elements, so for a multi-viewport case you must
specify unique elements for all viewports
* You may omit the explicit element (again, defaulting to the canvas)
for _one_ viewport in a multi-viewport definition
* The following properties/methods of the `Idetik` object ("runtime")
have been removed, and must be accessed via the viewport instead:
  * `camera`
  * `cameraControls`
  * `clientToClip`
  * `events`
  * `layerManager`
* You can access a viewport through the runtime object with:
* `viewports` property, which returns a readonly array of `Viewport`
objects
* `getViewport(id: string): Viewport | undefined` if you have a viewport
ID
* In the common case of a single-viewport app, use
`runtime.viewports[0]` to get the viewport
  
I also included a comprehensive JSDoc comment for the `Idetik`
constructor to document how it works. Hopefully this is helpful as we
look into generating API docs.

Finally, this includes a new example demonstrating multiple viewports in
`packages/core/examples/multi_viewport`. Because the `ChunkManager` and
associated machinery is not compatible with multiple viewports, I used a
`VolumeLayer` and `PerspectiveCamera` in the second viewport as a
temporary workaround. This example can serve as a testing rig as we add
multi-viewport capability to the chunking infrastructure.
<img width="977" height="838" alt="Screenshot 2025-10-22 at 10 21 33 AM"
src="https://github.com/user-attachments/assets/91ed89da-0d38-4f36-94aa-2216a3079112"
/>

### Related Issue
#486 (does not close)

### Tests & Checks
Tested existing examples and updated automated tests
Included a new example demonstrating multiple viewports, with the
limitation that only one can use the `ChunkManager`

N.B. I did not update the React code here in anticipation of #499
merging soon. We will need to update the react library accordingly when
this is published. Tests on this branch will probably fail until
incorporating the react package removal.

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>