Add bounding boxes to renderable objects (#513)
Summary:
This PR adds bounding boxes to renderable objects. These will be used for
frustum-culling debugging or any renderer-level culling we may need, though the main
path is more likely to generate boxes directly from chunks. The computation is lazy,
so it introduces no extra perf cost, but we can roll some of this work back if it's no
longer useful beyond debugging.