Fix some prefetched chunks from being disposed (#524)
### Summary
This fixes a bug that was introduced in #493 because `chunk.prefetch` is
not equivalent to `eligibleForPrefetch` in the `shouldDispose` method.

That bug can be reproduced on `main` by loading the default chunk
streaming example, panning the camera, and observing the chunk stats
changing in a way that they shouldn't (i.e. some prefetched chunks get
reloaded).

### Tests & Checks
I tested the chunk streaming example.