feat: add volume shaders for different texture data types (#525)
### Summary
<!---
  Please try to cover the following questions in your summary:
  1. Describe the product (or technical) problem you are solving.
  2. Explain the engineering solution you have chosen.
  3. Discuss the implementation choices/tradeoffs you have made.
-->

Primary goal here is to add support for different texture data types in
volume rendering.

This adds support for passing width, height, depth, and a Texture3D to
the volume renderable object. The dataType in the Texture3D is used to
determine the correct shader program name to use for the volume
renderable based on the type of data stored in the texture. This is
using a similar pattern to how the image renderable uses the texture to
determine the program name.

Three new shader programs are added, a `floatVolume`, `intVolume` and
`uintVolume`. Also added is an example volume rendering fragment shader,
and passing UVs from the vertex shader for volume rendering to the
fragment shader for volume rendering. This volume rendering is still
straight down Z of the volume, will be updated later to respect camera
parameters and allow inputs to modify normalization of values.

For the example of trying these, a simple checker board style small
texture has been defined in the volume layer.
https://github.com/chanzuckerberg/idetik/pull/515/files has more
progress on using actual data here, but this PR is separated into two to
hopefully be easier to review.

<img width="889" height="721" alt="image"
src="https://github.com/user-attachments/assets/a4bc97b6-316a-4623-88fe-d295887086f0"
/>


### Tests & Checks
<!---
  How did you validate that your changes were implemented correctly?
  
  Please think about the following prompts:
  1. I wrote automated tests.
  2. I manually tested my changes, and here's what I did:
-->

Manually tested changes by changing the data type of the texture used in
the volume layer example. Also checked other pre-existing examples just
in case. This changes the multiple-viewports example to look as:
<img width="1658" height="880" alt="image"
src="https://github.com/user-attachments/assets/de62ac3d-f097-47ac-83fa-8a2a2e3c435c"
/>